Split admin and user layouts with custom user navigation
